requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic
jinja2
deepdiff
python-dotenv
